add isrbn

+++ contracts/interfaces/ISRBN.sol
@@ -4,6 +4,7 @@ pragma solidity =0.8.4;
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface ISRBN is IERC20 {
-    function burn(uint256 amount) external;
-    function mint(address dst, uint rawAmount) external;
+  function burn(uint256 amount) external;
+
+  function mint(address dst, uint256 rawAmount) external;
 }
